integrate blueprintjs to account application
