--- apps/oauth2/priv/oauth2-account/package.json
@@ -1,25 +1,0 @@
-{
-  "name": "oauth2-account",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "dress-code": "^2.2.0",
-    "query-string": "^5.0.1",
-    "react": "^16.1.1",
-    "react-dom": "^16.1.1",
-    "react-dress-code": "git+https://git@github.com/fogfish/react-dress-code.git",
-    "react-redux": "^5.0.6",
-    "react-router-dom": "^4.2.2",
-    "react-scripts": "1.0.17",
-    "recompose": "^0.26.0",
-    "redux": "^3.7.2",
-    "redux-logger": "^3.0.6",
-    "redux-thunk": "^2.2.0"
-  },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test --env=jsdom",
-    "eject": "react-scripts eject"
-  }
-}
