--- apps/oauth2/priv/oauth2-account/public/redirect.html
@@ -1,117 +1,0 @@
-<!DOCTYPE html>
-<!--
-
-  https://tools.ietf.org/html/rfc6749#section-4.1.1
-
-  3.1.2.5.  Endpoint Content
-
-     The redirection request to the client's endpoint typically results in
-     an HTML document response, processed by the user-agent.  If the HTML
-     response is served directly as the result of the redirection request,
-     any script included in the HTML document will execute with full
-     access to the redirection URI and the credentials it contains.
-
-     The client SHOULD NOT include any third-party scripts (e.g., third-
-     party analytics, social plug-ins, ad networks) in the redirection
-     endpoint response.  Instead, it SHOULD extract the credentials from
-     the URI and redirect the user-agent again to another endpoint without
-     exposing the credentials (in the URI or elsewhere).  If third-party
-     scripts are included, the client MUST ensure that its own scripts
-     (used to extract and remove the credentials from the URI) will
-    execute first.
--->
-<meta charset="utf-8">
-<script type="text/javascript">
-
-//
-//
-var config = {
-   // application root endpoint
-   application: '/oauth2/account',
-
-   //
-   client_id: 'oauth2-account',
-
-   // oauth2 token end-point
-   token: '/oauth2/token'
-}
-
-//
-function encode(json)
-{ 
-   return Object.keys(json).map( 
-      function(key){return [encodeURIComponent(key), encodeURIComponent(json[key])].join('=')}
-   ).join('&')
-}
-
-//
-function decode(text)
-{
-   var keyval = function(key, val){return key==="" ? val : decodeURIComponent(val)}
-   return text 
-      ? JSON.parse('{"' + text.replace(/&/g, '","').replace(/=/g,'":"') + '"}', keyval)
-      : {}
-}
-
-//
-function redirect(url, json)
-{
-   window.location.replace(url + '?' + encode(json))
-}
-
-
-//
-function onError(json)
-{
-   redirect(config.application, {error: json.error})
-}
-
-//
-function onToken(json)
-{
-   var now = +new Date()
-   window.localStorage.setItem('access_token', 'Bearer ' + json.access_token)
-   window.localStorage.setItem('access_token_ttl', now + json.expires_in * 1000 - 60000)
-   redirect(config.application, {})
-}
-
-//
-function onCode(json)
-{
-   var xhr = new XMLHttpRequest();
-   xhr.open('POST', config.token);
-   xhr.setRequestHeader('Content-Type', 'application/x-www-form-urlencoded');
-   xhr.send(encode({
-      grant_type: 'authorization_code',
-      client_id: config.client_id,
-      code: json.code
-   }))
-   xhr.onreadystatechange = function() 
-   {
-      var DONE = 4
-      var OK = 200
-      if (xhr.readyState === DONE)
-      {
-         if (xhr.status === OK)
-         {
-            onToken(JSON.parse(xhr.responseText))
-         } else {
-            onError({error: 'access_denied'})
-         }
-      }
-   }
-}
-
-var result = decode(window.location.search.substring(1))
-if (result.error)
-{
-   onError(result)
-} else if (result.access_token) {
-   onToken(result)
-} else if (result.code) {
-   onCode(result)
-} else {
-   onError({error: "unsupported_response_type"})
-}
-
-</script>
