--- apps/oauth2/priv/oauth2-account/src/components/Account/Menu.js
@@ -1,11 +1,0 @@
-import React from 'react'
-import { Tab, TabElement } from 'react-dress-code'
-
-const Menu = () => (
-  <Tab vertical>
-    <TabElement header>Developer Settings</TabElement>
-    <TabElement active>OAuth Apps</TabElement>
-  </Tab>
-)
-
-export default Menu
