+++ uxis/account/src/app.css
@@ -1,4 +1,5 @@
 @import "../node_modules/dress-code/dist/css/dress-code.min.css";
+@import "../node_modules/@blueprintjs/core/lib/css/blueprint.css";
 
 .dc-row
 {
