+++ uxis/account/src/components/NewApp/Registrar/Endpoint.js
