+++ uxis/account/src/components/NewApp/Registrar/Registrar.js
