+++ uxis/account/src/components/NewApp/Registrar/index.js
