+++ uxis/account/src/index.js
@@ -4,7 +4,7 @@ import store from './store'
 import React from 'react'
 import { Provider } from 'react-redux'
 import ReactDOM from 'react-dom'
-import App from './App'
+import App from './app'
 
 
 ReactDOM.render(
